First batch of changes to permit VS2017 ARM64 tools to build required elements.

Related work items: #14268301

+++ ArmPkg/Drivers/CpuDxe/CpuDxe.c
@@ -224,7 +224,7 @@ InitializeDma (
   IN OUT  EFI_CPU_ARCH_PROTOCOL  *CpuArchProtocol
   )
 {
-  CpuArchProtocol->DmaBufferAlignment = ArmCacheWritebackGranule ();
+  CpuArchProtocol->DmaBufferAlignment = (UINT32)ArmCacheWritebackGranule ();
 }
 
 EFI_STATUS
